Add Two Numbers
I try to not use any extra space, so I mainly use the space of list 1 to store the result.
But when list 1 has iterate til end, I'll link it to list 2 and continue the iteration.

The total running time will be O(n) where n equal to the length of the longer list