Reverse Linked List
Use one pointer to point to the node that will be iterated next loop, and one pointer point to the temporary head of the finished list. An extra pointer handling the process.
So that will make the whole method being done by only one iteration through the whole list.

The total run time will be O(n)