767. Reorganize String
The first part is to determine if it is impossible to constuct a string that the neighbor are different characters.

If one character has more than half the length of the size of the string, then it's impossible.

Second part to construct the string. First fill up the string by fill in the blank in every two block then it will be over.

The total running time is O(n)