31. Next Permutation
First we want the decreasing pair from the back.
Then we swap the small number in the pair with the number that is just bigger to it and in it's rightward part. With this, we can make sure our new result is bigger.
But we still have to make it just a little bit bigger, so we have to sort the rightward part in increasing order.

The total running time would be O(n)