681. Next Closest Time
First, I test if there is any possible to get a minute that is hire than the original string.
If not, I'll test if I can get a combination to get an higher hour.

If both not, it means that it goes to next day, then we simply search for the lowest time of combination we got